moved liquidation_price method to exchange classes
