+++ freqtrade/leverage/__init__.py
@@ -1,3 +1,2 @@
 # flake8: noqa: F401
 from freqtrade.leverage.interest import interest
-from freqtrade.leverage.liquidation_price import liquidation_price
